Add sqlx-types feature for using enums directly
More work is needed either on this side, or more likely in the users of
this crate to avoid saving `None` or other NULL-like data. It would
help if we could somehow rename/alias variants as an Option::None
direction, but I'm not sure that's possible.

See https://github.com/ED-NEWP/galos_db/issues/1